Fix bug in bpfloader runtime restart
When netd restarted, it will run bpfloader again to make sure all
bpf programs are loaded properly. However, the bpfloader treat the
case program already loaded as a failure since it cannot get a fd for
it. It doesn't break the current implementation but it will show
unecessary error logs so it should be fixed.

Bug: 79938294
Test: runtime restart netd and no bpfloader error showed in logcat
Change-Id: Id8542cf6b68e67065e7aa9398b40d28beac4ecf9