Initial libnetd_resolv apex file
The goal about apex is that facilitate the update of system components
such as native services, HALs, runtime (ART), class libraries, etc.
that don’t fit into the standard Android application model.
Currently apex is still developing. For DNS module updatability,
we seperate dns related function(libnetd_resolv.so) from libc.so and
plan to use apex for it.This is not built by default.

Bug: 119527674
Test: m com.android.resolv; make; flash; adb shell ls /apex/
Change-Id: I5a273b2455a6d0be409377e81e2e13824181af27