make libjsoncpp be imported by only netd
netd_benchmark doesn't use log feature of aidl.
but because it statically imports aidl_interface(that turns on genlog),
it should import libjsoncpp.

when the patch in aidl is applied, it doesn't need to import libjsoncpp
in the case that module statically imports it.

so remove libjson import in netd_default, and add it into only netd

https://android-review.googlesource.com/c/platform/system/tools/aidl/+/909474
https://android-review.googlesource.com/c/platform/system/netd/+/853730/11/Android.bp#35

Bug: 124871490
Test: system/netd/tests/runtests.sh
Change-Id: Ibdc14fcc339134c9e22d074fa676f00729257f04