Ensure that addresses have actually been created in TunInterface.
Currently TunInterface assumes that IP addresses are created
immediately as soon as ifc_add_address returns. However, IPv6
addresses, even if created with IFA_F_NODAD or optimistic DAD,
are not created immediately and are not actually usable when the
netlink ACK returns.

Ensure tests aren't flaky by waiting for RTM_NEWADDR for the
address after creating it.

Test: system/netd/tests/runtests.sh
Change-Id: I5db5e25b24329ecbe953b592997b6c1358722c5c