Add back networking eBPF programs in netd
Since the networking eBPF programs are changed more often with other
netd userspace changes, It makes more sense to have the eBPF program in
netd project instead of in the bpf project. In the future we won't need
to submit separate CLs for both project. This is a direct copy of files
without modification.

Bug: 112334572
Test: dumpsys netd trafficcontroller
Change-Id: I50808dc7d62e62e8f343161d25167c315e801dcd