Remove netd_aidl_parcelables.
This was a stable AIDL interface that could be used in the
core framework which could not depend on the entire netd stable
AIDL interface (otherwise the definitions of these classes in
frameworks.jar would override the stable AIDL classes themselves,
making the networkstack app unable to use them).

The framework no longer depends on these classes, so remove the
build target.

Test: m
Bug: 128804404
Change-Id: I3ce6267d511588d4bc3d31123d29331c0181d629