binder_test: Avoid extra std::string copies
When a function argument is a const std::string reference, it's
inefficient to pass a std::string::c_str().  We're not particularly
concerned about the performance of this test, but by fixing up
this issue, it makes using static analysis to find this issue in
our code base less noisy.

Test: TreeHugger
Change-Id: I69fefcaa480ca6b0106e3362821acd73052668b9