change group permission of bpf maps
some bpf maps were only allowed to be accessed by
processes with root. But vendor process can't run
with root so that they are forbidden to set up any
ipatbles rule under the chain which has bpf maps
rules.

So the group permission of bpf maps need to be
changed to NET_ADMIN and allow process with NET_ADMIN
to read bpf map.

BUG: 129650054
Test: netdutils_test netd_unit_test netd_integration_test
Change-Id: Ibe39e9a7227910516a7ddb14d3d22f19904d6598