ebpf Clat - rename to ClatIngress throughout
The specific fields stored in the key and value structs necessitates
the use of this map only for ingress clat ipv6 -> ipv4 decoding purposes.

At some point we will want to implement ebpf clat egress, and then we'll
need new key/value pairs (and map).

Even though current implementation of eBPF program is ingress only,
it is a bad idea to name these fields src and dst since it can
quickly become confusing.  It is much easier to think of things as
remote and local or simply call something a prefix.

Whether 'local' and 'remote' are 'src' or 'dst' depends on traffic
flow direction...

Test: atest netd_unit_test
Bug: 65674744
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Iaf4185a547f8dbf8f59ade286eb0fb4b8d09f368