Fix the unit test for stats swap
after merging uid stats map and tagStatsMap to one map and implementing
the stats swap mechanism. The trafficController unit test needs to be
slighty modified to correctly reflect the behavior change of
trafficController. Also, dup the fake map fd for mock controller so
the fake map object and the mock controller map object no longer share
the same fd underneath. It fixes the double close problem in the
unit test.

Bug: 130280727
Test: netd_unit_test
Change-Id: I5534d2fd783da3b38d26359c9d62b2427a1725da