Fix flaky test BinderTest#NetworkRejectNonSecureVpn
Previously, we cache the all ip rule when test start, and compare
cached rules and system rules to ensure that rules is not changed.
However, it is possible that netd is called to modify rules during the test.
(e.g. networkDestroy() as below). Then the test will fail.
Remove the comparison of rules in the test to solve this problem.

Error log:
05-27 20:19:02.470 networkCreatePhysical(100, 0) <0.10ms>
05-27 20:19:02.495 networkAddInterface(100, "rmnet_data2") <1.91ms>
05-27 20:19:11.389 networkRejectNonSecureVpn("true") <0.35ms>
05-27 20:19:11.457 networkDestroy(100) <3.15ms>  <-- change ip rule during this test
05-27 20:19:11.466 networkRejectNonSecureVpn("false") <0.34ms>
05-27 20:19:11.539 networkRejectNonSecureVpn("false") -> ServiceSpecificException(2, "No such file or directory") <0.71ms>

Bug: 133818065
Test: cd system/netd && atest
Change-Id: Ic226f305a30e1cfa3b3b3f07a25987688b869578