Remove resolv/AndroidTest.xml
The presence of this file confuses atest into running
netd_integration_test instead of resolv_unit_test.

This file was added in aosp/720345 to increase the per-test timeout to
120 seconds for resolv_integration_test due to the following tests being
slower in CI than in local runs:

 - ResolverTest#GetAddrInfoStressTest_Binder_100
 - ResolverTest#GetAddrInfoStressTest_Binder_100000

But this might no longer be necessary, it seems that tests execute in a
few seconds in current TreeHugger runs.

Test: run treehugger
Change-Id: Iffd2e6fa0d7ae7af0dc1707ded6b24970eb18ee2