Restore the PrivateDnsMode emun to its original definition
Westworld requires the first field of all enums to be set to
UNKNOWN/UNSET, but this is awkward for the resolver code and
makes it hard to verify whether we have any PrivateDnsMode
variables initialized to UNKNOWN by accident.

So revert this change, and add a simple function to convert
the internal PrivateDnsMode enum to the other PrivateDnsMode
emun defined in the stats proto.

Bug: 119862317
Test: atest
Change-Id: Id146a16e41e0183867b3a4e2d87a2b0fc09df27d