Don't create sp<> to stack allocated mHwSvc
RefBase-derived objects should always be allocated in the heap.
They should also immediately become the target of an sp.

The original NetdHwService did not allow correct use. It privately
(and indirectly) inherited from RefBase, making it impossible to
generate an sp<> to it, as all clients should. We change the interface
to make this possible.

Bug: 138956784
Test: Booted with change.
Change-Id: Ia977f8a7f1a8bb554fc5e0bfcce056939c688812