Use Linux default replay window for IPsec
Replay window size allows for improved of an IPsec SA over L2 links
that may experience out-of-order delivery.

Relaxing the replay window size does NOT impact the security guarantees
provided by IPsec, as it still rejects replayed packets. If an attacker
has the keys to generate the older packets, they would likewise have the
keys to generate newer packets.

Impact of increasing the replay window size is primarily the memory
usage required, and thus should not be increased too high.

The Linux kernel uses 32 by default, and without a strong reason to
clamp this down to 4 (which would make the SA more lossy), we should
maintain the Linux default.

Bug: 142967324
Test: Tests passing
Change-Id: Iabebbf139ab73e52a9b8c9367f585f105d58689d