Disable preupload check for missing "Test:" header
Presubmit tests have been there for a while now, there's no need
to add noise in the commitlog to assert that, yes, we ran the usual
tests in the usual way.

Of course, when a change requires special manual testing steps, we can
still add a useful description instead of the unhelpful incantation
"Test: manually" that many have used to get around this check!

(Note the prominent absence of a Test: header here, which constitutes
the self-test for this change).

Change-Id: I1626816a7c03b4714c23363d04b717872182d460