Enable pedantic checks for uninitialized variables
The -Wsometimes-uninitialized and -Wconditional-uninitialized flags are
not enabled with -Wall because they lead to some false positives.
However, these have become exceedingly rare. I found only two such cases
in the netd codebase and a single one in DnsResolver.

Access to uninitialzied memory is one of the nastiest forms of undefined
behavior in C, it's time to eradicate it.

Change-Id: Iea5cf450e0e3c1f2443c06b1fac670290abe2e64