Support getting fwmark for a network
NetworkStack will use the tcp info queried from kernel to
diagnose internet health. The diagnosis should only focus on
specific network, e.g. default network. NetworkStack needs a way
to filter the target network. The only identifier is the fwmark
value for each socket. The fwmark calculation may be modified
in native and may not sync with NetworkStack. Thus, NetworkStack
will need a way to get the netId mask and the network fwmark to
know the network information contained in the fwmark. Expose this
function to ensure the fwmark implementation is aligned wih netd.

Bug: 130325409
Test: cd system/netd; atest
Change-Id: I52fba39e041490016224beffb273693e64ce4338