share eBPF struct definitions between ebpf and C++ netd
This is the main commit in a set of 3 commits across 3 diff git repos.
The other two are:
  https://android-review.googlesource.com/c/platform/system/bpf/+/1199649
  "remove network specific struct definitions from BpfUtils.h"
and
  https://android-review.googlesource.com/c/platform/frameworks/base/+/1200738
  "minor change to keep it building"

We move the struct definitions to bpf_shared.h so they can be
shared between C++ netd and C ebpf code.

They also become typedefs and are renamed for more consistent naming.
(there's some weird issue with ebpf compiler on some devices with
non-typedef'ed structs)

Test: builds, atest
Bug: 146787904
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I324c0ab9db0186dcea0ec9ee33140909be285bc4