Change group owner and permission of tethering eBPF map for network stack
The default tethering eBPF map was only allowed to be accessed by
processes with root. But network stack process doesn't run with
root so it is not able to access the map.

So the group owner and permission of the map need to be changed to
NETWORK_STACK and allow network stack process to read/write the map.

Test: build and check that both group owner and the group permission of
the maps are as expected, network_stack and group read/write.

$ adb shell ls -l /sys/fs/bpf | egrep -e "offload(.*)map"
-rw-rw---- 1 root network_stack [..] map_offload_tether_ingress_map

Change-Id: I6627d01a5e2998a298dbd40215d45e3b6e776276