Move attaching and detaching clsact from ClatdController to RouteController
The clsact attaching and detaching control plane is moved to:
  RouteController::addInterfaceToPhysicalNetwork      - add clsact
  RouteController::removeInterfaceFromPhysicalNetwork - del clsact

The above change implies that the clsact lifetime for each interface
has been extended from clat enabled time to interface lifetime.

The only exception is that attaching clsact to v4- tun interface
still lives in ClatdController. The reason is that clat is started
before the v4- tun interface is added to the network and clat has
already needed to add the bpf filters.

After all, keep attaching and detaching clat {in, e}gress bpf
filters in ClatdController.

Test: manual clatd test
1. Connect to IPv6-Only WiFi hotspot
2. Browse 172.217.0.46 (google.com) successfully
3. Disconnect from WiFi
Repeat the above steps three times.

Change-Id: I971e105484c7678ac304788e5ffff7cc709c400d